backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
langchain>=0.0.200
google-generativeai>=0.3.0
pypdf>=3.15.0
chromadb>=0.4.10
python-dotenv>=1.0.0         